config: add option to remove warning about git clean

+++ cli/src/commands/git/init.rs
@@ -133,7 +133,7 @@ pub fn cmd_git_init(
         r#"Initialized repo in "{}""#,
         relative_wc_path.display()
     )?;
-    if colocate {
+    if colocate && command.settings().git_settings()?.warn_about_clean {
         writeln!(
             ui.hint_default(),
             r"Running `git clean -xdf` will remove `.jj/`!",
